requirements.txt
aircv==1.4.6
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
exceptiongroup==1.2.1
flatbuffers==24.3.25
h11==0.14.0
humanfriendly==10.0
idna==3.7
iniconfig==2.0.0
more-itertools==10.2.0
mpmath==1.3.0
numpy==1.26.4
onnxruntime==1.18.0
opencv-python-headless==4.9.0.80
outcome==1.3.0.post0
packaging==24.0
pillow==10.3.0
pluggy==0.13.1
protobuf==5.27.0
py==1.11.0
PySocks==1.7.1
pytest==7.1.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
selenium==3.141.0
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.12.1
toml==0.10.2
tomli==2.0.1
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.0
urllib3==1.26.2
wcwidth==0.2.13
webdriver-manager==4.0.1
websockets==12.0
wsproto==1.2.0
